% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{diff_months}
\alias{diff_months}
\title{Difference in months between dates}
\usage{
diff_months(x, y)
}
\arguments{
\item{x}{integer date string}

\item{y}{integer date string}
}
\description{
Difference in months between dates
}
