% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{diff_years}
\alias{diff_years}
\title{Difference in fraction of years between dates}
\usage{
diff_years(x, y)
}
\arguments{
\item{x}{integer date string}

\item{y}{integer date string}
}
\description{
Difference in fraction of years between dates
}
