% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{is_between}
\alias{is_between}
\title{Check if a date is between two other dates}
\usage{
is_between(start, end, x)
}
\arguments{
\item{start}{start date}

\item{end}{end date}

\item{x}{point date}
}
\description{
Check if a date is between two other dates
}
