% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{overlapping}
\alias{overlapping}
\title{Check if episodes overlapp}
\usage{
overlapping(start, end, start2, end2)
}
\arguments{
\item{start}{of episode 1}

\item{end}{of eposide 2}

\item{start2}{of episode 2}

\item{end2}{of episode 2}
}
\description{
Check if episodes overlapp
}
