% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_fert_data.R
\name{to_fert_data}
\alias{to_fert_data}
\title{Transfrom ehd_data to fertility data}
\usage{
to_fert_data(x)
}
\arguments{
\item{x}{ehd_data}
}
\description{
Starts 1 year before marriage
ends at end
remove all events between conseption and birth
}
